requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
scikit-learn==1.5.1
scipy==1.13.1


# NLP / Retrieval
sentence-transformers==3.0.1
rank-bm25==0.2.2
faiss-cpu==1.8.0.post1


# Ranking
lightgbm==4.3.0


# Training / Logs
torch==2.3.1
transformers==4.43.3
loguru==0.7.2
transformers[torch]


# Testing
pytest==8.2.1
pytest-cov==5.0.0


# Utils
tqdm==4.66.4